Renamed the platform modules.
